+++ include/roq/oms/order_update.hpp
@@ -17,7 +17,6 @@ struct ROQ_PUBLIC OrderUpdate final {
   OrderType order_type = {};
   TimeInForce time_in_force = {};
   Mask<ExecutionInstruction> execution_instructions;
-  std::string_view order_template;
   std::chrono::nanoseconds create_time_utc = {};
   std::chrono::nanoseconds update_time_utc = {};
   std::string_view external_account;
@@ -59,7 +58,6 @@ struct fmt::formatter<roq::oms::OrderUpdate> {
         R"(order_type={}, )"
         R"(time_in_force={}, )"
         R"(execution_instructions={}, )"
-        R"(order_template="{}", )"
         R"(create_time_utc={}, )"
         R"(update_time_utc={}, )"
         R"(external_account="{}", )"
@@ -85,7 +83,6 @@ struct fmt::formatter<roq::oms::OrderUpdate> {
         value.order_type,
         value.time_in_force,
         value.execution_instructions,
-        value.order_template,
         value.create_time_utc,
         value.update_time_utc,
         value.external_account,