+++ include/roq/order_update.hpp
@@ -42,7 +42,6 @@ struct ROQ_PUBLIC OrderUpdate final {
   OrderType order_type = {};                          //!< Order type
   TimeInForce time_in_force = {};                     //!< Time in force
   Mask<ExecutionInstruction> execution_instructions;  //!< Execution instructions
-  std::string_view order_template;                    //!< Order template
   std::chrono::nanoseconds create_time_utc = {};      //!< Created timestamp (UTC)
   std::chrono::nanoseconds update_time_utc = {};      //!< Updated timestamp (UTC)
   std::string_view external_account;                  //!< External account name
@@ -97,7 +96,6 @@ struct fmt::formatter<roq::OrderUpdate> {
         R"(order_type={}, )"
         R"(time_in_force={}, )"
         R"(execution_instructions={}, )"
-        R"(order_template="{}", )"
         R"(create_time_utc={}, )"
         R"(update_time_utc={}, )"
         R"(external_account="{}", )"
@@ -130,7 +128,6 @@ struct fmt::formatter<roq::OrderUpdate> {
         value.order_type,
         value.time_in_force,
         value.execution_instructions,
-        value.order_template,
         value.create_time_utc,
         value.update_time_utc,
         value.external_account,